# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TMJConnect API
  version: 1.0.0
  description: |
    HIPAA-compliant orofacial pain management platform API.
    Supports patients, providers, and admin roles.
  contact:
    name: TMJConnect Engineering
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PatientProfile:
      description: Patient user profile with demographic information
      type: object
      required: [id, email, role, is_active, created_at, first_name, last_name, timezone]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        phone: { type: [string, "null"] }
        role: { type: string, enum: [patient] }
        is_active: { type: boolean }
        mfa_enabled: { type: boolean, description: "Whether patient has opted into MFA" }
        created_at: { type: string, format: date-time }
        first_name: { type: string }
        last_name: { type: string }
        date_of_birth: { type: [string, "null"], format: date }
        gender: { type: [string, "null"] }
        avatar_url: { type: [string, "null"], format: uri }
        city: { type: [string, "null"] }
        state: { type: [string, "null"] }
        timezone: { type: string, description: "IANA timezone (e.g. America/New_York)" }
...
    PageMeta:
      description: Offset-based pagination metadata
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, description: "Current page number" }
        limit: { type: integer, description: "Items per page" }
        total: { type: integer, description: "Total matching items" }
        totalPages: { type: integer, description: "Total pages" }
...
paths:
...
  /patients/me:
    get:
      tags: [Patients]
      summary: Get current patient profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/PatientProfile' }
    patch:
      tags: [Patients]
      summary: Update patient profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name: { type: string }
                last_name: { type: string }
                date_of_birth: { type: [string, "null"], format: date }
                gender: { type: [string, "null"] }
                city: { type: [string, "null"] }
                state: { type: [string, "null"] }
                timezone: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/PatientProfile' }
    delete:
      tags: [Patients]
      summary: Request account deletion (soft delete)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Account deletion requested }
  /patients/me/sessions:
    get:
      tags: [Patients]
      summary: List active sessions
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Session' }
  /patients/me/sessions/{sessionId}:
    delete:
      tags: [Patients]
      summary: Revoke a session
      security: [{ bearerAuth: [] }]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Session revoked }
...
  /providers/me:
    get:
      tags: [Providers]
      summary: Get current provider profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ProviderProfile' }
    patch:
      tags: [Providers]
      summary: Update provider profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ProviderProfile' }
  /providers/patients:
    get:
      tags: [Providers]
      summary: List linked patients with 7-day stats
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: search
          in: query
          schema: { type: string }
          description: Search by first or last name
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PatientDashboardItem' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
  /providers/patients/{patientId}:
    get:
      tags: [Providers]
      summary: Get linked patient detail
      security: [{ bearerAuth: [] }]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Patient detail }
        '403': { description: Patient not linked to provider }
  /providers/patients/{patientId}/assignments:
    get:
      tags: [Providers]
      summary: List a patient's exercise assignments
      security: [{ bearerAuth: [] }]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ExerciseAssignment' }
    post:
      tags: [Providers]
      summary: Assign an exercise to a patient
      security: [{ bearerAuth: [] }]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [exercise_id]
              properties:
                exercise_id: { type: string, format: uuid }
                frequency: { type: string, default: daily }
                sets: { type: integer, minimum: 1, default: 1 }
      responses:
        '201': { description: Assignment created }
        '403': { description: Patient not linked }
  /providers/assignments/{assignmentId}:
    patch:
      tags: [Providers]
      summary: Update an exercise assignment
      security: [{ bearerAuth: [] }]
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                frequency: { type: string }
                sets: { type: integer, minimum: 1 }
                status: { type: string, enum: [active, paused, completed] }
      responses:
        '200': { description: Assignment updated }
    delete:
      tags: [Providers]
      summary: Delete an exercise assignment
      security: [{ bearerAuth: [] }]
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Assignment deleted }
  /providers/exercises:
    get:
      tags: [Providers]
      summary: List provider's exercise library
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Exercise' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
    post:
      tags: [Providers]
      summary: Create an exercise in the library
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 255 }
                description: { type: [string, "null"] }
                duration_seconds: { type: [integer, "null"] }
                category: { type: [string, "null"] }
                instructions: { type: [string, "null"] }
                video_url: { type: [string, "null"], format: uri }
                thumbnail_url: { type: [string, "null"], format: uri }
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Exercise' }
  /providers/exercises/{exerciseId}:
    patch:
      tags: [Providers]
      summary: Update an exercise
      security: [{ bearerAuth: [] }]
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Exercise updated }
    delete:
      tags: [Providers]
      summary: Delete an exercise
      security: [{ bearerAuth: [] }]
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Exercise deleted (cascades to assignments) }

  # ──────────────────────────── REPORTS ────────────────────────────
...
  /providers/patients/{patientId}/notes:
    get:
      tags: [Providers]
      summary: List clinical notes for a patient (provider-scoped; never patient-visible)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ClinicalNote' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
    post:
      tags: [Providers]
      summary: Create a clinical note (provider-only)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, maxLength: 10000 }
                tags:
                  type: array
                  maxItems: 20
                  items: { type: string, maxLength: 50 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ClinicalNote' }
  /providers/notes/{noteId}:
    patch:
      tags: [Providers]
      summary: Update a clinical note
      security: [{ bearerAuth: [] }]
      parameters:
        - name: noteId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body: { type: string, maxLength: 10000 }
                tags:
                  type: array
                  items: { type: string, maxLength: 50 }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ClinicalNote' }
    delete:
      tags: [Providers]
      summary: Delete a clinical note
      security: [{ bearerAuth: [] }]
      parameters:
        - name: noteId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
  /providers/patients/{patientId}/report-requests:
    get:
      tags: [Providers]
      summary: List report requests sent to this patient (provider-scoped)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ReportRequest' }
    post:
      tags: [Providers]
      summary: Request a report from a patient
      description: Persists a pending request and enqueues a `report_requested` notification.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string, maxLength: 2000 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ReportRequest' }
  /providers/dashboard/summary:
    get:
      tags: [Providers]
      summary: Dashboard aggregate (single round-trip replacing 4 list calls)
      description: Returns counts + recent patients + urgent inbox. Frontend dashboard hits this instead of fanning out.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      activePatients: { type: integer }
                      unreadReports: { type: integer }
                      pendingCodes: { type: integer }
                      urgentReports: { type: integer }
                      recentPatients:
                        type: array
                        items:
                          type: object
                          properties:
                            patient_id: { type: string, format: uuid }
                            first_name: { type: string }
                            last_name: { type: string }
                            last_symptom_at: { type: [string, "null"], format: date-time }
                      urgentInbox:
                        type: array
                        items: { $ref: '#/components/schemas/ReportInboxItem' }
...
  /patients/me/activity:
    get:
      tags: [Patients]
      summary: Get account activity log
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ActivityEvent' }

  # ──────────────────────── TRACKING — JAW MOBILITY ─────────────────────────
...